changed information app and migrations made
