+++ acm_web_site/apps/business/web_site/admin.py
@@ -5,9 +5,7 @@ from django.core.files.storage import FileSystemStorage
 from .models import *
 from .forms import (
 	GalleryAdminForm,
-	AwardAdminForm,
 	ActivityAdminForm,
-	MemberAdminForm,
 	ProjectAdminForm,
 	TeamAdminForm,
 	TutorialAdminForm
@@ -25,19 +23,6 @@ class GalleryAdmin(admin.ModelAdmin):
 		obj.picture = uploaded_file_url
 		super().save_model(request, obj, form, change)
 
-
-class AwardAdmin(admin.ModelAdmin):
-	form = AwardAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
 class ActivityAdmin(admin.ModelAdmin):
 	form = ActivityAdminForm
 
@@ -48,19 +33,6 @@ class ActivityAdmin(admin.ModelAdmin):
 		uploaded_file_url = fs.url(filename)
 		obj.poster = uploaded_file_url
 		super().save_model(request, obj, form, change)
-
-
-class MemberAdmin(admin.ModelAdmin):
-	form = MemberAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
 
 class ProjectAdmin(admin.ModelAdmin):
 	form = ProjectAdminForm
@@ -99,9 +71,7 @@ class TutorialAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Contest)
-admin.site.register(Member, MemberAdmin)
 admin.site.register(Team, TeamAdmin)
-admin.site.register(Award, AwardAdmin)
 admin.site.register(Gallery, GalleryAdmin)
 admin.site.register(Tutorial, TutorialAdmin)
 admin.site.register(Activity, ActivityAdmin)
